+++ Robust.Client/Replays/Loading/ReplayLoadManager.Read.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
-using Robust.Shared.ContentPack;
 using Robust.Shared.GameStates;
 using Robust.Shared.Serialization;
 using Robust.Shared.Serialization.Markdown;
@@ -13,15 +12,17 @@ using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 using Robust.Shared.Replays;
-using static Robust.Shared.Replays.IReplayRecordingManager;
+using static Robust.Shared.Replays.ReplayConstants;
 
 namespace Robust.Client.Replays.Loading;
 
 public sealed partial class ReplayLoadManager
 {
     [SuppressMessage("ReSharper", "UseAwaitUsing")]
-    public async Task<ReplayData> LoadReplayAsync(IWritableDirProvider dir, ResPath path, LoadReplayCallback callback)
+    public async Task<ReplayData> LoadReplayAsync(IReplayFileReader fileReader, LoadReplayCallback callback)
     {
+        using var _ = fileReader;
+
         if (_client.RunLevel == ClientRunLevel.Initialize)
             _client.StartSinglePlayer();
         else if (_client.RunLevel != ClientRunLevel.SinglePlayerGame)
@@ -32,31 +33,21 @@ public sealed partial class ReplayLoadManager
         List<ReplayMessage> messages = new();
 
         var compressionContext = new ZStdCompressionContext();
-        var metaData = LoadMetadata(dir, path);
+        var metaData = LoadMetadata(fileReader);
 
-        // Strip tailing "/"
-        // why is there no method for this.
-        if (path.CanonPath.EndsWith("/"))
-            path = new(path.CanonPath.Substring(0, path.CanonPath.Length - 1));
-
-        var total = dir.Find($"{path.ToRelativePath()}/*.{Ext}").files.Count();
-
-        // Exclude string & init event files from the total.
-        total--;
-        if (dir.Exists(path / InitFile))
-            total--;
+        var totalData = fileReader.AllFiles.Count(x => x.Filename.StartsWith(DataFilePrefix));
 
         var i = 0;
         var intBuf = new byte[4];
-        var name = path / $"{i++}.{Ext}";
-        while (dir.Exists(name))
+        var name = new ResPath($"{DataFilePrefix}{i++}.{Ext}");
+        while (fileReader.Exists(name))
         {
-            await callback(i+1, total, LoadingState.ReadingFiles, false);
+            await callback(i+1, totalData, LoadingState.ReadingFiles, false);
 
-            using var fileStream = dir.OpenRead(name);
+            using var fileStream = fileReader.Open(name);
             using var decompressStream = new ZStdDecompressStream(fileStream, false);
 
-            fileStream.Read(intBuf);
+            fileStream.ReadExactly(intBuf);
             var uncompressedSize = BitConverter.ToInt32(intBuf);
 
             var decompressedStream = new MemoryStream(uncompressedSize);
@@ -72,12 +63,16 @@ public sealed partial class ReplayLoadManager
                 messages.Add(msg);
             }
 
-            name = path / $"{i++}.{Ext}";
+            name = new ResPath($"{DataFilePrefix}{i++}.{Ext}");
         }
-        DebugTools.Assert(i - 1 == total);
-        await callback(total, total, LoadingState.ReadingFiles, false);
 
-        var initData = LoadInitFile(dir, path, compressionContext);
+        // Could happen if there's gaps in the numbers of the data.
+        if (i - 1 != totalData)
+            throw new Exception("Could not read expected amount of data files from replay");
+
+        await callback(totalData, totalData, LoadingState.ReadingFiles, false);
+
+        var initData = LoadInitFile(fileReader, compressionContext);
         compressionContext.Dispose();
 
         var (checkpoints, serverTime) = await GenerateCheckpointsAsync(
@@ -101,54 +96,67 @@ public sealed partial class ReplayLoadManager
     }
 
     private ReplayMessage? LoadInitFile(
-        IWritableDirProvider dir,
-        ResPath path,
+        IReplayFileReader fileReader,
         ZStdCompressionContext compressionContext)
     {
-        if (!dir.Exists(path / InitFile))
+        if (!fileReader.Exists(FileInit))
             return null;
 
         // TODO replays compress init messages, then decompress them here.
-        using var fileStream = dir.OpenRead(path / InitFile);
+        using var fileStream = fileReader.Open(FileInit);
         _serializer.DeserializeDirect(fileStream, out ReplayMessage initData);
         return initData;
     }
 
-    public MappingDataNode? LoadYamlMetadata(IWritableDirProvider directory, ResPath resPath)
+    public MappingDataNode? LoadYamlMetadata(IReplayFileReader fileReader)
     {
-        if (!directory.Exists(resPath / MetaFile))
+        return LoadYamlFile(fileReader, FileMeta);
+    }
+
+    public MappingDataNode? LoadYamlFinalMetadata(IReplayFileReader fileReader)
+    {
+        return LoadYamlFile(fileReader, FileMetaFinal);
+    }
+
+    private static MappingDataNode? LoadYamlFile(IReplayFileReader fileReader, ResPath path)
+    {
+        if (!fileReader.Exists(path))
             return null;
 
-        using var file = directory.OpenRead(resPath / MetaFile);
+        using var file = fileReader.Open(path);
         var parsed = DataNodeParser.ParseYamlStream(new StreamReader(file));
         return parsed.FirstOrDefault()?.Root as MappingDataNode;
     }
 
     private (MappingDataNode YamlData, HashSet<string> CVars, TimeSpan Duration, TimeSpan StartTime, bool ClientSide)
-        LoadMetadata(IWritableDirProvider directory, ResPath path)
+        LoadMetadata(IReplayFileReader fileReader)
     {
         _sawmill.Info($"Reading replay metadata");
-        var data = LoadYamlMetadata(directory, path);
+        var data = LoadYamlMetadata(fileReader);
         if (data == null)
-            throw new Exception("Failed to parse yaml metadata");
+            throw new Exception("Failed to load yaml metadata");
 
-        var typeHash = Convert.FromHexString(((ValueDataNode) data[Hash]).Value);
-        var stringHash = Convert.FromHexString(((ValueDataNode) data[Strings]).Value);
-        var startTick = ((ValueDataNode) data[Tick]).Value;
-        var timeBaseTick = ((ValueDataNode) data[BaseTick]).Value;
-        var timeBaseTimespan = ((ValueDataNode) data[BaseTime]).Value;
-        var clientSide = bool.Parse(((ValueDataNode) data[IsClient]).Value);
-        var duration = TimeSpan.Parse(((ValueDataNode) data[Duration]).Value);
+        var finalData = LoadYamlFinalMetadata(fileReader);
+        if (finalData == null)
+            throw new Exception("Failed to load final yaml metadata");
+
+        var typeHash = Convert.FromHexString(((ValueDataNode) data[MetaKeyTypeHash]).Value);
+        var stringHash = Convert.FromHexString(((ValueDataNode) data[MetaKeyStringHash]).Value);
+        var startTick = ((ValueDataNode) data[MetaKeyStartTick]).Value;
+        var timeBaseTick = ((ValueDataNode) data[MetaKeyBaseTick]).Value;
+        var timeBaseTimespan = ((ValueDataNode) data[MetaKeyBaseTime]).Value;
+        var clientSide = bool.Parse(((ValueDataNode) data[MetaKeyIsClientRecording]).Value);
+        var duration = TimeSpan.Parse(((ValueDataNode) finalData[MetaFinalKeyDuration]).Value);
 
         if (!typeHash.SequenceEqual(_serializer.GetSerializableTypesHash()))
             throw new Exception($"{nameof(IRobustSerializer)} hashes do not match. Loading replays using a bad replay-client version?");
 
-        using var stringFile = directory.OpenRead(path / StringsFile);
+        using var stringFile = fileReader.Open(FileStrings);
         var stringData = new byte[stringFile.Length];
-        stringFile.Read(stringData);
+        stringFile.ReadExactly(stringData);
         _serializer.SetStringSerializerPackage(stringHash, stringData);
 
-        using var cvarsFile = directory.OpenRead(path / CvarFile);
+        using var cvarsFile = fileReader.Open(FileCvars);
         // Note, this does not invoke the received-initial-cvars event. But at least currently, that doesn't matter
         var cvars = _confMan.LoadFromTomlStream(cvarsFile);